Set up Continuous Integration
For pull requests, this tries to build the container (which
also runs unit tests, due to how the Dockerfile is set up).
GitHub will displays the build status on the review page for
the pull request.

For changes to the main branch, this builds the container
and pushes it to the GitHub container registry.